+++ src/users.js
@@ -13,166 +13,6 @@ var FS = require('./FamilySearch'),
 
 /**
  * @ngdoc function
- * @name user.types:constructor.User
- * @description
- *
- * User - a user is returned from {@link user.functions:getCurrentUser getCurrentUser};
- * Contributor Ids are agent ids, not user ids.
- */
-var User = function() {
-
-};
-
-User.prototype = {
-  constructor: User
-  /**
-   * @ngdoc property
-   * @name user.types:constructor.User#id
-   * @propertyOf user.types:constructor.User
-   * @return {String} Id of the user
-   */
-
-  /**
-   * @ngdoc property
-   * @name user.types:constructor.User#personId
-   * @propertyOf user.types:constructor.User
-   * @return {String} id of the {@link person.types:constructor.Person Person} for this user
-   */
-
-  /**
-   * @ngdoc property
-   * @name user.types:constructor.User#treeUserId
-   * @propertyOf user.types:constructor.User
-   * @return {String} agent (contributor) id of the user
-   */
-
-  /**
-   * @ngdoc property
-   * @name user.types:constructor.User#contactName
-   * @propertyOf user.types:constructor.User
-   * @return {String} contact name
-   */
-
-  /**
-   * @ngdoc property
-   * @name user.types:constructor.User#displayName
-   * @propertyOf user.types:constructor.User
-   * @return {String} full display name
-   */
-
-  /**
-   * @ngdoc property
-   * @name user.types:constructor.User#givenName
-   * @propertyOf user.types:constructor.User
-   * @return {String} given name
-   */
-
-  /**
-   * @ngdoc property
-   * @name user.types:constructor.User#familyName
-   * @propertyOf user.types:constructor.User
-   * @return {String} family name
-   */
-
-  /**
-   * @ngdoc property
-   * @name user.types:constructor.User#gender
-   * @propertyOf user.types:constructor.User
-   * @return {String} MALE or FEMALE
-   */
-
-  /**
-   * @ngdoc property
-   * @name user.types:constructor.User#email
-   * @propertyOf user.types:constructor.User
-   * @return {String} email address
-   */
-
-  /**
-   * @ngdoc property
-   * @name user.types:constructor.User#preferredLanguage
-   * @propertyOf user.types:constructor.User
-   * @return {String} e.g., en
-   */
-};
-
-/**
- * @ngdoc function
- * @name user.types:constructor.Agent
- * @description
- *
- * An agent is returned from {@link user.functions:getAgent getAgent}.
- * Contributor Ids are agent ids, not user ids.
- */
-var Agent = function() {
-
-};
-
-Agent.prototype = {
-  constructor: Agent,
-  /**
-   * @ngdoc property
-   * @name user.types:constructor.Agent#id
-   * @propertyOf user.types:constructor.Agent
-   * @return {String} Id of the agent
-   */
-
-  /**
-   * @ngdoc function
-   * @name user.types:constructor.Agent#$getName
-   * @methodOf user.types:constructor.Agent
-   * @function
-   * @return {String} name of the agent
-   */
-  $getName: function() { return maybe(maybe(this.names)[0]).value; },
-
-  /**
-   * @ngdoc function
-   * @name user.types:constructor.Agent#$getAccountName
-   * @methodOf user.types:constructor.Agent
-   * @function
-   * @return {String} account / contact name of the agent
-   */
-  $getAccountName: function() { return maybe(maybe(this.accounts)[0]).accountName; },
-
-
-  /**
-   * @ngdoc function
-   * @name user.types:constructor.Agent#$getEmail
-   * @methodOf user.types:constructor.Agent
-   * @function
-   * @return {String} email of the agent
-   */
-  $getEmail: function() {
-    var email = maybe(maybe(this.emails)[0]).resource;
-    return email ? email.replace(/^mailto:/,'') : email;
-  },
-
-  /**
-   * @ngdoc function
-   * @name user.types:constructor.Agent#$getPhoneNumber
-   * @methodOf user.types:constructor.Agent
-   * @function
-   * @return {String} phone number of the agent
-   */
-  $getPhoneNumber: function() {
-    return maybe(maybe(this.phones)[0]).resource;
-  },
-
-  /**
-   * @ngdoc function
-   * @name user.types:constructor.Agent#$getAddress
-   * @methodOf user.types:constructor.Agent
-   * @function
-   * @return {String} postal address of the agent
-   */
-  $getAddress: function() {
-    return maybe(maybe(this.addresses)[0]).value;
-  }
-};
-
-/**
- * @ngdoc function
  * @name user.functions:getCurrentUser
  * @function
  *
@@ -197,7 +37,7 @@ FS.prototype.getCurrentUser = function(params, opts) {
       return self.plumbing.get(url, params, {}, opts,
         utils.compose(
           utils.objectExtender({getUser: function() { return maybe(this.users)[0]; }}),
-          utils.constructorSetter(User, 'users')
+          utils.constructorSetter(FS.User, 'users')
         ));
     });
 };
@@ -229,7 +69,7 @@ FS.prototype.getAgent = function(aid, params, opts) {
       return self.plumbing.get(url, params, {}, opts,
         utils.compose(
           utils.objectExtender({getAgent: function() { return maybe(this.agents)[0]; }}),
-          utils.constructorSetter(Agent, 'agents')
+          utils.constructorSetter(FS.Agent, 'agents')
         ));
     });
 };
@@ -260,6 +100,3 @@ FS.prototype.getMultiAgent = function(aids, params, opts) {
   });
   return self.helpers.promiseAll(promises);
 };
-
-FS.Agent = Agent;
-FS.User = User;
